+++ resources/assets/js/components/Administradores/Asignaciones/Credencial/index.vue
