+++ resources/assets/js/components/Profesores/Notas/NotaComponent.vue
@@ -1,248 +1,205 @@
 <template>
-   <v-flex>
-      <combostudents
-         :students="students"
-         :dialogOpens="dialogOpen"
-         v-on:speak="selectMethod($event)"
-      ></combostudents>
-      <v-flex md8 xs12 offset-md2>
-         <v-layout wrap row>
-            <v-flex xs12 pt-2 px-3>
-               <v-toolbar dark color="green" class="py-1">
+   <v-form ref="form">
+      <v-layout wrap mx-5 xs6>
+         <v-flex xs6 pt-2 px-3>
+            <v-text-field
+               xs12
+               v-model="nota.titulo"
+               :rules="[rules.required, rules.between]"
+               counter
+               label="*Titulo"
+               name="titulo"
+               required
+            ></v-text-field>
+         </v-flex>
+         <v-flex xs6 pt-2 px-3>
+            <v-text-field
+               xs12
+               v-model="nota.valor_porcentual"
+               :rules="[rules.required]"
+               label="*Valor Porcentual"
+               name="valor_porcentual"
+               suffix="%"
+               :mask="mask"
+               :return-masked-value="true"
+               required
+            ></v-text-field>
+         </v-flex>
+         <v-flex xs6 pt-2 px-3>
+            <v-text-field
+               xs12
+               v-model="nota.porcentaje_obtenido"
+               :rules="[rules.required]"
+               label="*Porcentaje Obtenido"
+               name="porcentaje_obtenido"
+               suffix="%"
+               :mask="mask"
+               :return-masked-value="true"
+               required
+            ></v-text-field>
+         </v-flex>
+         <v-flex xs6 pt-2 px-3>
+            <v-text-field
+               xs12
+               v-model="nota.descripcion"
+               counter
+               label="Descripción"
+               name="descripcion"
+            ></v-text-field>
+         </v-flex>
+      </v-layout>
+      <v-layout wrap mx-5 xs6>
+         <v-flex xs6 pt-2 px-3>
+            <v-menu
+               ref="menu"
+               v-model="menu"
+               :close-on-content-click="false"
+               :nudge-right="40"
+               :return-value.sync="nota.fecha"
+               lazy
+               transition="scale-transition"
+               offset-y
+               full-width
+               min-width="290px"
+            >
+               <template v-slot:activator="{ on }">
+                  <v-text-field
+                     v-model="nota.fecha"
+                     label="*Fecha"
+                     prepend-icon="event"
+                     readonly
+                     v-on="on"
+                  ></v-text-field>
+               </template>
+               <v-date-picker v-model="nota.fecha" no-title scrollable>
                   <v-spacer></v-spacer>
-                  <v-toolbar-title>Listado de Calificaciones</v-toolbar-title>
-                  <v-spacer></v-spacer>
-               </v-toolbar>
-               <v-card>
-                  <dialogcreate v-on:speak="crearMethod($event)"></dialogcreate>
-                  <dialogdelete
-                     :dialogDeletes="dialogDelete"
-                     :text="text"
-                     :title="title"
-                     v-on:speak="borrarMethod($event)"
-                  ></dialogdelete>
-                  <dialogedit
-                     :dialogEdits="dialogEdit"
-                     :editNota="nota"
-                     v-on:speak="editarMethod($event)"
-                  ></dialogedit>
-                  <v-card-title>
-                     <v-spacer></v-spacer>Notas
-                     <v-spacer></v-spacer>
-                     <v-text-field
-                        v-model="search"
-                        append-icon="search"
-                        label="Buscar por Titulo"
-                        single-line
-                        hide-details
-                     ></v-text-field>
-                  </v-card-title>
-                  <v-data-table
-                     :headers="headers"
-                     :items="notas"
-                     :search="search"
-                     rowsPerPageText="Elementos por página:"
-                     rowsPerPageAll="Todos"
-                     pageText="{0}-{1} de {2}"
-                     noResultsText="Ningún resultado a mostrar"
-                     nextPage="Página siguiente"
-                     prevPage="Página anterior"
-                     noDataText="Ningún dato disponible"
-                  >
-                     <template v-slot:items="props">
-                        <td
-                           class="text-xs-left"
-                           :style="{backgroundColor: (props.index % 2 == 0) ?'#c8e6c9' : 'white'}"
-                        >{{ props.item.titulo }}</td>
-                        <td
-                           class="text-xs-center"
-                           :style="{backgroundColor: (props.index % 2 == 0) ?'#c8e6c9' : 'white'}"
-                        >Realizada: {{ props.item.fecha }} - Trimestre {{ props.item.trimestre }}</td>
-                        <td
-                           class="text-xs-center"
-                           :style="{backgroundColor: (props.index % 2 == 0) ?'#c8e6c9' : 'white'}"
-                        >Valor: {{ props.item.valor_porcentual }} - Obtenido: {{ props.item.porcentaje_obtenido }}</td>
-                        <td
-                           class="text-xs-center"
-                           :style="{backgroundColor: (props.index % 2 == 0) ?'#c8e6c9' : 'white'}"
-                        >{{ props.item.nota }}</td>
-                        <td
-                           class="text-xs-center"
-                           :style="{backgroundColor: (props.index % 2 == 0) ?'#c8e6c9' : 'white'}"
-                        >
-                           <v-btn
-                              flat
-                              small
-                              color="primary"
-                              @click="buscar(props.item.id),dialogEdit = true"
-                           >Editar</v-btn>
-                        </td>
-                        <td
-                           class="text-xs-center"
-                           :style="{backgroundColor: (props.index % 2 == 0) ?'#c8e6c9' : 'white'}"
-                        >
-                           <v-btn
-                              flat
-                              small
-                              color="error"
-                              dark
-                              @click.stop="buscar(props.item.id),dialogDelete = true"
-                           >Eliminar</v-btn>
-                        </td>
-                     </template>
-                     <v-alert
-                        v-slot:no-results
-                        :value="true"
-                        color="error"
-                        icon="warning"
-                     >Tu busqueda por "{{ search }}" no encontro resultados.</v-alert>
-                  </v-data-table>
-               </v-card>
-            </v-flex>
-         </v-layout>
-      </v-flex>
-   </v-flex>
+                  <v-btn flat color="primary" @click="menu = false">Cancel</v-btn>
+                  <v-btn flat color="primary" @click="$refs.menu.save(nota.fecha)">OK</v-btn>
+               </v-date-picker>
+            </v-menu>
+         </v-flex>
+         <v-flex xs6 pt-2 px-3>
+            <v-select
+               :items="tipos"
+               :rules="[rules.required]"
+               item-text="text"
+               item-value="value"
+               :menu-props="{returnValue:'value'}"
+               v-model="nota.tipo"
+               label="*Tipo"
+               d-block
+            ></v-select>
+         </v-flex>
+         <v-flex xs6 pt-2 px-3>
+            <v-select
+               :items="condiciones"
+               :rules="[rules.required]"
+               item-text="text"
+               item-value="value"
+               :menu-props="{returnValue:'value'}"
+               v-model="nota.condicion"
+               label="*Condición"
+               d-block
+            ></v-select>
+         </v-flex>
+         <v-flex xs6 pt-2 px-3>
+            <v-select
+               :items="trimestres"
+               :rules="[rules.required]"
+               item-text="text"
+               item-value="value"
+               :menu-props="{returnValue:'value'}"
+               v-model="nota.trimestre"
+               label="*Trimestre"
+               d-block
+            ></v-select>
+         </v-flex>
+      </v-layout>
+   </v-form>
 </template>
 
 <script>
-import combostudents from "../../Modals/ComboStudents.vue";
-import dialogdelete from "../../Modals/DialogDelete.vue";
-import dialogcreate from "./DialogCreate.vue";
-import dialogedit from "./DialogEdit.vue";
-
 export default {
-   components: {
-      combostudents,
-      dialogcreate,
-      dialogedit,
-      dialogdelete
-   },
+   name: "nota_component",
+   props: ["validar", "restaurar", "datos"],
    data() {
       return {
-         search: "",
-         dialogOpen: false,
-         dialogDelete: false,
-         dialogEdit: false,
-         title: "Eliminar Nota?",
-         text: "Seguro que quieres eliminar esta Nota.",
-         notas: [],
-         nota: {
-            id: null
+         menu: false,
+         rules: {
+            required: value => !!value || "Requerido.",
+            between: v =>
+               (!!v && (v.length >= 3 && v.length < 51)) ||
+               "Entre 3 y 50 Caracteres"
          },
-         students: [],
-         student: null,
-         mensaje: null,
-         headers: [
+         mask: "##.##",
+         trimestres: [
+            { text: "Primer Trimestre", value: 1 },
+            { text: "Segundo Trimestre", value: 2 },
+            { text: "Tercer Trimestre", value: 3 }
+         ],
+         condiciones: [
+            { text: "Aprobada", value: "Aprobada" },
+            { text: "Reprobada", value: "Reprobada" },
+            { text: "Realisada", value: "Realisada" },
+            { text: "No realisada", value: "No realisada" }
+         ],
+         tipos: [
+            { text: "Examen", value: "Examen" },
+            { text: "Tarea", value: "Tarea" },
             {
-               text: "Titulo:",
-               align: "left",
-               sortable: false,
-               value: "titulo"
+               text: "Trabajo o investigación",
+               value: "Trabajo o investigación"
             },
-            { text: "Fecha:", align: "center", value: "fecha" },
-            {
-               text: "Porcentajes:",
-               align: "center",
-               value: "valor_porcentual"
-            },
-            { text: "Nota:", align: "center", value: "nota" },
-            { text: "Editar", align: "center", value: "" },
-            { text: "Eliminar", align: "center", value: "" }
-         ]
+            { text: "Otra", value: "Otra" }
+         ],
+         nota: {
+            id: null,
+            titulo: "",
+            valor_porcentual: "",
+            porcentaje_obtenido: "",
+            tipo: "",
+            condicion: "",
+            descripcion: "",
+            trimestre: "",
+            fecha: new Date().toISOString().substr(0, 10),
+            student_id: null,
+            course_id: null,
+            estado: 1
+         }
       };
    },
-   methods: {
-      selectMethod: function(msj) {
-         if (msj.Selected && msj.Student != null) {
-            this.student = msj.Student;
-            this.reset();
-         } else this.$router.push("screenteacher");
-      },
-      buscar(id) {
-         this.nota = this.notas.find(function(c) {
-            return c.id == id;
-         });
-      },
-      crearMethod: function(msg) {
-         var self = this;
-         if (msg.titulo != "") {
-            msg.student_id = this.student.id;
-            axios.post("qualifications/", msg).then(function(res) {
-               self.mensaje = res.data;
-               self.reset();
-            });
-         }
-      },
-      reset() {
-         var self = this;
-         axios
-            .get("/qualificationsforstudent/" + this.student.id)
-            .then(function(res) {
-               let qualifications = res.data;
-               qualifications.forEach(qualification => {
-                  let x =
-                     (100 / qualification.valor_porcentual) *
-                     qualification.porcentaje_obtenido;
-                  qualification.nota = x.toFixed(2);
-               });
-               self.notas = qualifications;
-               self.nota = null;
-            });
-      },
-      eliminar() {
-         var self = this;
-         if (this.nota.id != null) {
-            axios.delete("qualifications/" + this.nota.id).then(function(res) {
-               self.mensaje = res.data;
-               self.reset();
-            });
-         }
-      },
-      editar() {
-         var self = this;
-         if (this.nota.id != null) {
-            axios
-               .put("qualifications/" + this.nota.id, this.nota)
-               .then(function(res) {
-                  self.mensaje = res.data;
-                  self.reset();
-               });
-         }
-      },
-      borrarMethod: function(msg) {
-         this.dialogDelete = msg.dialogdelete;
-         if (msg.Delete) this.eliminar();
-      },
-      editarMethod: function(msg) {
-         this.dialogEdit = msg.dialogedit;
-         if (msg.Edit) {
-            this.nota = msg.nota;
-            this.editar();
-         } else {
-            this.nota = null;
-         }
-      }
-   },
    watch: {
-      mensaje(val) {
-         if (val.type === "success")
-            this.$toast.success(val.message, {
-               y: "top",
-               timeout: 6000
-            });
-         else
-            this.$toast.error(val.message, {
-               y: "top",
-               timeout: 6000
-            });
+      validar(val) {
+         if (val) {
+            if (this.$refs.form.validate()) {
+               this.$emit("speak", Object.assign({}, this.nota));
+               return;
+            }
+         }
+         this.$emit("speak", null);
+      },
+      datos(val) {
+         if (val != null) {
+            this.nota.id = val.id;
+            this.nota.titulo = val.titulo;
+            this.nota.valor_porcentual = val.valor_porcentual;
+            this.nota.porcentaje_obtenido = val.porcentaje_obtenido;
+            this.nota.tipo = val.tipo;
+            this.nota.condicion = val.condicion;
+            this.nota.descripcion = val.descripcion;
+            this.nota.trimestre = val.trimestre;
+            this.nota.fecha = val.fecha;
+            this.nota.student_id = val.student_id;
+            this.nota.course_id = val.course_id;
+            this.nota.estado = val.estado;
+         }
+      },
+      restaurar(val) {
+         if (val) this.$refs.form.reset();
       }
-   },
-   mounted() {
-      let self = this;
-      axios.get("/studentsforcourse").then(function(res) {
-         self.students = res.data;
-         self.dialogOpen = true;
-      });
    }
 };
 </script>
+
+<style scoped>
+</style>
