+++ resources/assets/js/components/Profesores/Noticias/index.vue
