+++ resources/assets/js/router.js
@@ -2,22 +2,22 @@
 import Vue from 'vue';
 import Router from 'vue-router';
 import Profesorcomponent from './components/Administradores/Profesor/index.vue';
-import Profesornotacomponent from './components/Profesores/Notas/NotaComponent.vue';
+import Profesornotacomponent from './components/Profesores/Notas/index.vue';
 import Profesornotasagregarcomponent from './components/Profesores/Notas/NotasAgregarComponent.vue';
-import Profesorpromediocomponent from './components/Profesores/Promedios/PromedioComponent.vue';
-import Profesortareacomponent from './components/Profesores/Tareas/TareaComponent.vue';
+import Profesorpromediocomponent from './components/Profesores/Promedios/index.vue';
+import Profesortareacomponent from './components/Profesores/Tareas/index.vue';
 import Estudiantecomponent from './components/Administradores/Estudiante/index.vue';
 import Estudiantecursoscomponent from './components/Estudiantes/CursosComponent.vue';
 import Horariocomponent from './components/Administradores/Horario/index.vue';
 import Seccioncomponent from './components/Administradores/Seccion/index.vue';
 import Asignaturacomponent from './components/Administradores/Asignatura/index.vue';
-import Credencialcomponent from './components/Administradores/Asignaciones/Credencial.vue';
+import Credencialcomponent from './components/Administradores/Asignaciones/Credencial/index.vue';
 import Cursocomponent from './components/Administradores/Curso/index.vue';
-import Admincomponent from './components/Administradores/AdministracionDefault.vue';
-import AsigSeccionAlumnoscomponent from './components/Administradores/Asignaciones/SeccionAlumnos.vue';
-import AsigCursoLeccionescomponent from './components/Administradores/Asignaciones/CursoLecciones.vue';
+import Admincomponent from './components/Administradores/index.vue';
+import AsigSeccionAlumnoscomponent from './components/Administradores/Asignaciones/SeccionAlumno/index.vue';
+import AsigCursoLeccionescomponent from './components/Administradores/Asignaciones/CursoLeccion/index.vue';
 import Administradornoticiascomponent from './components/Administradores/Noticias/index.vue';
-import Profesornoticiascomponent from './components/Profesores/Noticias/NewsComponent.vue';
+import Profesornoticiascomponent from './components/Profesores/Noticias/index.vue';
 import Estudiantenoticiasglobalcomponent from './components/Estudiantes/Noticias/NewsGlobalComponent.vue';
 import Estudiantenoticiasgrupalcomponent from './components/Estudiantes/Noticias/NewsGrupalComponent.vue';
 import Estudiantenotacomponent from './components/Estudiantes/QualificationsComponent.vue';
